Move `query_archetype_with_history` to make it more globally available
Small refactor discussed with @Wumpf to make `query_archetype_with_history` consistent with `latest_at_with_blueprint_resolved_data`
