+++ crates/viewer/re_space_view/src/query.rs
@@ -1,12 +1,12 @@
 use nohash_hasher::IntSet;
 
-use re_chunk_store::{LatestAtQuery, RangeQuery, RowId};
-use re_log_types::TimeInt;
-use re_query::LatestAtResults;
-use re_types_core::ComponentName;
-use re_viewer_context::{DataResult, ViewContext, ViewerContext};
-
 use crate::results_ext::{HybridLatestAtResults, HybridRangeResults};
+use crate::HybridResults;
+use re_chunk_store::{LatestAtQuery, RangeQuery, RowId};
+use re_log_types::{TimeInt, Timeline};
+use re_query::LatestAtResults;
+use re_types_core::{Archetype, ComponentName};
+use re_viewer_context::{DataResult, QueryRange, ViewContext, ViewerContext};
 
 // ---
 
@@ -122,6 +122,48 @@ pub fn latest_at_with_blueprint_resolved_data<'a>(
     }
 }
 
+pub fn query_archetype_with_history<'a>(
+    ctx: &'a ViewContext<'a>,
+    timeline: &Timeline,
+    timeline_cursor: TimeInt,
+    query_range: &QueryRange,
+    component_names: impl IntoIterator<Item = ComponentName>,
+    data_result: &'a re_viewer_context::DataResult,
+) -> HybridResults<'a> {
+    match query_range {
+        QueryRange::TimeRange(time_range) => {
+            let range_query = RangeQuery::new(
+                *timeline,
+                re_log_types::ResolvedTimeRange::from_relative_time_range(
+                    time_range,
+                    timeline_cursor,
+                ),
+            );
+            let results = range_with_blueprint_resolved_data(
+                ctx,
+                None,
+                &range_query,
+                data_result,
+                component_names,
+            );
+            (range_query, results).into()
+        }
+        QueryRange::LatestAt => {
+            let latest_query = LatestAtQuery::new(*timeline, timeline_cursor);
+            let query_shadowed_defaults = false;
+            let results = latest_at_with_blueprint_resolved_data(
+                ctx,
+                None,
+                &latest_query,
+                data_result,
+                component_names,
+                query_shadowed_defaults,
+            );
+            (latest_query, results).into()
+        }
+    }
+}
+
 fn query_overrides<'a>(
     ctx: &ViewerContext<'_>,
     data_result: &re_viewer_context::DataResult,
@@ -194,6 +236,14 @@ pub trait DataResultQuery {
         latest_at_query: &'a LatestAtQuery,
     ) -> HybridLatestAtResults<'a>;
 
+    fn query_archetype_with_history<'a, A: re_types_core::Archetype>(
+        &'a self,
+        ctx: &'a ViewContext<'a>,
+        timeline: &Timeline,
+        timeline_cursor: TimeInt,
+        query_range: &QueryRange,
+    ) -> HybridResults<'a>;
+
     fn best_fallback_for<'a>(
         &self,
         ctx: &'a ViewContext<'a>,
@@ -215,6 +265,23 @@ impl DataResultQuery for DataResult {
             self,
             A::all_components().iter().copied(),
             query_shadowed_defaults,
+        )
+    }
+
+    fn query_archetype_with_history<'a, A: Archetype>(
+        &'a self,
+        ctx: &'a ViewContext<'a>,
+        timeline: &Timeline,
+        timeline_cursor: TimeInt,
+        query_range: &QueryRange,
+    ) -> HybridResults<'a> {
+        query_archetype_with_history(
+            ctx,
+            timeline,
+            timeline_cursor,
+            query_range,
+            A::all_components().iter().copied(),
+            self,
         )
     }
